refactor mock to add common objects
