+++ build.gradle
@@ -5,6 +5,7 @@ buildscript {
   repositories {
     mavenCentral()
     jcenter()
+    mavenLocal()
   }
   dependencies {
     classpath("org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}")
@@ -43,6 +44,7 @@ sourceCompatibility = 1.8
 targetCompatibility = 1.8
 
 repositories {
+  mavenLocal()
   mavenCentral()
   jcenter()
 }
@@ -66,6 +68,7 @@ dependencies {
   compile('org.springframework.boot:spring-boot-starter-web')
   compile('org.skyscreamer:jsonassert:1.3.0')
   compile('org.opensaml:opensaml:2.6.4')
+  compile('gov.ca.emsa.pulse:common:0.0.1-SNAPSHOT')
   testCompile('org.springframework.boot:spring-boot-starter-test')
 }
