+++ src/main/java/gov/ca/emsa/controller/DirectoryController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import gov.ca.emsa.Utils;
-import gov.ca.emsa.domain.Organization;
+import gov.ca.emsa.pulse.common.domain.Organization;
 import io.swagger.annotations.Api;
 
 @RestController
