--- src/main/java/gov/ca/emsa/domain/Document.java
@@ -1,26 +1,0 @@
-package gov.ca.emsa.domain;
-
-public class Document {
-	private String id;
-	private String name;
-	private Patient patient;
-	
-	public String getId() {
-		return id;
-	}
-	public void setId(String id) {
-		this.id = id;
-	}
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
-	public Patient getPatient() {
-		return patient;
-	}
-	public void setPatient(Patient patient) {
-		this.patient = patient;
-	}
-}
