+++ src/main/java/gov/ca/emsa/domain/Organizations.java
@@ -1,6 +1,10 @@
 package gov.ca.emsa.domain;
 
+
+import gov.ca.emsa.pulse.common.domain.Organization;
+
 import java.util.List;
+
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
